backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
aiohttp==3.9.5
apscheduler==3.10.4
